requirements.txt
Flask>=2.2,<3.0
numpy>=1.24
pandas>=2.0
opencv-python-headless>=4.8
Pillow>=10.0
pyheif>=0.7
fuzzywuzzy>=0.18
python-Levenshtein>=0.12
easyocr>=1.7
torch>=2.0
geopy>=2.4
requests>=2.31
scikit-learn>=1.2
